fix(release): treat merged milestone PRs as closed
